git: mailinfo
